feature: move some commands to Games cog
- Solve #52
- Add new message images
